+++ core/src/main/scala/kafka/server/KafkaApis.scala
@@ -2600,9 +2600,9 @@ class KafkaApis(val requestChannel: RequestChannel,
   }
 
   def handleAlterConfigsRequest(request: RequestChannel.Request): Unit = {
+    val alterConfigsRequest = request.body[AlterConfigsRequest]
     if (!request.isForwarded && config.usesSelfManagedQuorum) {
       // If using KRaft, per broker config alterations should be validated on the broker that the config(s) is for before forwarding them to the controller
-      val alterConfigsRequest = request.body[AlterConfigsRequest]
       val brokerConfigs = alterConfigsRequest.configs.asScala.filter(entry => entry._1.`type` == ConfigResource.Type.BROKER)
 
       // Validate per-broker dynamic configs
@@ -2656,7 +2656,6 @@ class KafkaApis(val requestChannel: RequestChannel,
       }
     } else if (config.requiresZookeeper) {
       val zkSupport = metadataSupport.requireZkOrThrow(KafkaApis.shouldAlwaysForward(request))
-      val alterConfigsRequest = request.body[AlterConfigsRequest]
       val (authorizedResources, unauthorizedResources) = alterConfigsRequest.configs.asScala.toMap.partition { case (resource, _) =>
         resource.`type` match {
           case ConfigResource.Type.BROKER_LOGGER =>
@@ -2785,47 +2784,66 @@ class KafkaApis(val requestChannel: RequestChannel,
   }
 
   def handleIncrementalAlterConfigsRequest(request: RequestChannel.Request): Unit = {
+    val alterConfigsRequest = request.body[IncrementalAlterConfigsRequest]
+    val configs = alterConfigsRequest.data.resources.iterator.asScala.map { alterConfigResource =>
+      val configResource = new ConfigResource(ConfigResource.Type.forId(alterConfigResource.resourceType),
+        alterConfigResource.resourceName)
+      configResource -> alterConfigResource.configs.iterator.asScala.map {
+        alterConfig => new AlterConfigOp(new ConfigEntry(alterConfig.name, alterConfig.value),
+          OpType.forId(alterConfig.configOperation))
+      }.toBuffer
+    }.toMap
+
     if (!request.isForwarded && config.usesSelfManagedQuorum) {
       // If using KRaft, per broker config alterations should be validated on the broker that the config(s) is for before forwarding them to the controller
-      //val alterConfigsRequest = request.body[IncrementalAlterConfigsRequest]
-      //val brokerConfigs = alterConfigsRequest.configs.asScala.filter(entry => entry._1.`type` == ConfigResource.Type.BROKER)
+      val results = configs.map { case (resource, alterConfigOps) =>
+        try {
+          if (resource.`type` == ConfigResource.Type.BROKER) {
+              // In ZK case, the old config is retrieved, altered then validated
+              // val persistentProps = if (perBrokerConfig) adminKRaftClient.fetchEntityConfig(ConfigType.Broker, brokerId.get.toString)
+              // else adminKRaftClient.fetchEntityConfig(ConfigType.Broker, ConfigEntityName.Default)
+              // val configProps = this.config.dynamicConfig.fromPersistentProps(persistentProps, perBrokerConfig)
 
-      //TODO: Validate dynamic config update
-      //
-      // In ZK cluster the following is done 
-      //
-      // For each config resource with BROKER (BROKER_LOGGER should also be done but is handled differently)
-      //
-      // val persistentProps = if (perBrokerConfig) adminZkClient.fetchEntityConfig(ConfigType.Broker, brokerId.get.toString)
-      // else adminZkClient.fetchEntityConfig(ConfigType.Broker, ConfigEntityName.Default)
-      // val configProps = this.config.dynamicConfig.fromPersistentProps(persistentProps, perBrokerConfig)
-      // prepareIncrementalConfigs(alterConfigOps, configProps, KafkaConfig.configKeys)
-      // alterBrokerConfigs(resource, alterConfigsRequest.validateOnly, configProps, configEntryMap)
-      //
-      // So in Kraft the current configs should also be fetched somehow before validating
-      // I do not think all of the dynamic configs needed are gauranteed to be in metadata cache so need a client to send describe request
-      //
-      // val persistentProps = if (perBrokerConfig) adminKRaftClient.fetchEntityConfig(ConfigType.Broker, brokerId.get.toString)
-      // else adminKRaftClient.fetchEntityConfig(ConfigType.Broker, ConfigEntityName.Default)
-      // val configProps = this.config.dynamicConfig.fromPersistentProps(persistentProps, perBrokerConfig)
-      // prepareIncrementalConfigs(alterConfigOps, configProps, KafkaConfig.configKeys)
-      // configHelper.validateBrokerConfigs(resource, alterConfigsRequest.validateOnly, configProps, configEntryMap)
+              // For now in the KRaft case, just validate new configs without retrieving the old config
+              val configProps = new Properties
 
-      // Now forward the request to the controller so that the configs are persisted to the metadata quorum.
-      // This will never fail since this request has not already been forwarded and this is a KRaft broker.
-      forwardToControllerOrFail(request)
+              val configEntriesMap = alterConfigOps.map(entry => (entry.configEntry.name, entry.configEntry.value)).toMap
+              configHelper.prepareIncrementalConfigs(alterConfigOps.toSeq, configProps, KafkaConfig.configKeys)
+              configHelper.validateBrokerConfigs(resource, alterConfigsRequest.data.validateOnly, configProps, configEntriesMap)
+          } else if (resource.`type` == ConfigResource.Type.BROKER_LOGGER) {
+              configHelper.getAndValidateBrokerId(resource)
+              configHelper.validateLogLevelConfigs(alterConfigOps.toSeq)
+              resource -> ApiError.NONE
+          } else {
+              resource -> ApiError.NONE
+          }
+        } catch {
+          case e @ (_: ConfigException | _: IllegalArgumentException) =>
+            val message = s"Invalid config value for resource $resource: ${e.getMessage}"
+            info(message)
+            resource -> ApiError.fromThrowable(new InvalidRequestException(message, e))
+          case e: Throwable =>
+            // Log client errors at a lower level than unexpected exceptions
+            val message = s"Error processing alter configs request for resource $resource, config $alterConfigOps"
+            if (e.isInstanceOf[ApiException])
+              info(message, e)
+            else
+              error(message, e)
+            resource -> ApiError.fromThrowable(e)
+        }
+      }.toMap
+
+      if (results.filterNot(_._2 == ApiError.NONE).nonEmpty || alterConfigsRequest.data.validateOnly) {
+        // If validation fails for any reason or if validateOnly is true, send response back to the client and don't forward the request to the controller.
+        requestHelper.sendResponseMaybeThrottle(request, requestThrottleMs => new IncrementalAlterConfigsResponse(
+          requestThrottleMs, results.asJava))
+      } else {
+        // Now forward the request to the controller so that the configs are authorized and persisted to metadata.
+        // This will never fail since this request has not already been forwarded and this is a KRaft broker.
+        forwardToControllerOrFail(request)
+      }
     } else if (config.requiresZookeeper) {
       val zkSupport = metadataSupport.requireZkOrThrow(KafkaApis.shouldAlwaysForward(request))
-      val alterConfigsRequest = request.body[IncrementalAlterConfigsRequest]
-
-      val configs = alterConfigsRequest.data.resources.iterator.asScala.map { alterConfigResource =>
-        val configResource = new ConfigResource(ConfigResource.Type.forId(alterConfigResource.resourceType),
-          alterConfigResource.resourceName)
-        configResource -> alterConfigResource.configs.iterator.asScala.map {
-          alterConfig => new AlterConfigOp(new ConfigEntry(alterConfig.name, alterConfig.value),
-            OpType.forId(alterConfig.configOperation))
-        }.toBuffer
-      }.toMap
 
       val (authorizedResources, unauthorizedResources) = configs.partition { case (resource, _) =>
         resource.`type` match {
